Move to editor folder
